hle: kernel: k_memory_layout: Move KMemoryRegionAllocator out of global.
